clang-format + clang-tidy: zero-warning baseline, fix real bugs
Tooling:
- Add .clang-format (LLVM base, 2-space/120-col, SortIncludes: Never)
- Add .clang-tidy (bugprone, clang-analyzer, performance, portability)
- scripts/format.sh (--check for CI), scripts/lint.sh
- CMake exports compile_commands.json
- Vendored argon2 excluded from formatting and linting

Whole-tree reformat (597 files, vendor excluded).

Config fixes:
- NTFS_DEVICE_OPERATIONS_IOCTL_INT -> _UL (fixes incompatible ioctl fn-pointer)
- Enable HAVE_UUID_GENERATE (GPT UUIDs now actually generated)

Bugs fixed:
- backup.c: use-after-free / heap corruption when fread() fails on
  cluster data (freed be/clist were still written to and linked into list).
  Also fixes be->path leak on error path.
- backup.c: model_len cap aligned to buffer size (was 512, buffer is 256)
- backup.c: misplace widening-cast on fseek fixed
- backup.c: removed dead format_size(), logged unused counters
- carver.c: cancelled flag now forces a final incomplete session save
  at cancel offset, and no longer fabricates a bogus trailing carve entry
- restore.c: removed redundant strdup/free and dead bytes_done counter

Warnings:
- sign-compare, pointer-sign, unused-var fixes across backup, jpg, pdb,
  tiff, ttf, txt, pdf, ps, partgptw
- -Wno-unused-parameter, -Wno-address-of-packed-member (architectural,
  documented in CMakeLists)
- Build exits with zero warnings

AGENTS.md: build/format/lint workflow and zero-warning policy.

+++ src/formats/file_ps.c
@@ -32,82 +32,68 @@
 #include "filegen.h"
 #include "common.h"
 
-
-
 static void register_header_check_ps(file_stat_t *file_stat);
 
-const file_hint_t file_hint_ps= {
-  .extension="ps",
-  .description="PostScript or Encapsulated PostScript document",
-  .max_filesize=PHOTOREC_MAX_FILE_SIZE,
-  .recover=1,
-  .enable_by_default=1,
-  .register_header_check=&register_header_check_ps
-};
+const file_hint_t file_hint_ps = {.extension = "ps",
+                                  .description = "PostScript or Encapsulated PostScript document",
+                                  .max_filesize = PHOTOREC_MAX_FILE_SIZE,
+                                  .recover = 1,
+                                  .enable_by_default = 1,
+                                  .register_header_check = &register_header_check_ps};
 
-static const unsigned char ps_header[11]= { '%','!','P','S','-','A','d','o','b','e','-'};
+static const unsigned char ps_header[11] = {'%', '!', 'P', 'S', '-', 'A', 'd', 'o', 'b', 'e', '-'};
 
-
-static void file_check_ps(file_recovery_t *file_recovery)
-{
-  const unsigned char ps_footer[5]="%%EOF";
+static void file_check_ps(file_recovery_t *file_recovery) {
+  const unsigned char ps_footer[5] = "%%EOF";
   file_search_footer(file_recovery, ps_footer, sizeof(ps_footer), 1);
 }
 
-
-static data_check_t data_check_ps(const unsigned char *buffer, const unsigned int buffer_size, file_recovery_t *file_recovery)
-{
+static data_check_t data_check_ps(const unsigned char *buffer, const unsigned int buffer_size,
+                                  file_recovery_t *file_recovery) {
   unsigned int i;
-  
-  
-  
-  for(i=(buffer_size/2)-4;i+4<buffer_size;i++)
-  {
-    if(buffer[i]=='%' && buffer[i+1]=='%' && buffer[i+2]=='E' && buffer[i+3]=='O' && buffer[i+4]=='F')
-    {
-      file_recovery->calculated_file_size=file_recovery->file_size+i+5-(buffer_size/2);
+
+  for (i = (buffer_size / 2) - 4; i + 4 < buffer_size; i++) {
+    if (buffer[i] == '%' && buffer[i + 1] == '%' && buffer[i + 2] == 'E' && buffer[i + 3] == 'O' &&
+        buffer[i + 4] == 'F') {
+      file_recovery->calculated_file_size = file_recovery->file_size + i + 5 - (buffer_size / 2);
       return DC_STOP;
     }
   }
-  file_recovery->calculated_file_size=file_recovery->file_size+(buffer_size/2);
+  file_recovery->calculated_file_size = file_recovery->file_size + (buffer_size / 2);
   return DC_CONTINUE;
 }
 
-
-static int header_check_ps(const unsigned char *buffer, const unsigned int buffer_size, const unsigned int safe_header_only, const file_recovery_t *file_recovery, file_recovery_t *file_recovery_new)
-{
+static int header_check_ps(const unsigned char *buffer, const unsigned int buffer_size,
+                           const unsigned int safe_header_only, const file_recovery_t *file_recovery,
+                           file_recovery_t *file_recovery_new) {
   /* PS or EPSF */
   int i;
   reset_file_recovery(file_recovery_new);
-  file_recovery_new->min_filesize=sizeof(ps_header);
-  file_recovery_new->file_check=&file_check_ps;
-  
-  for(i=sizeof(ps_header); i < 20; i++)
-  {
-    switch(buffer[i])
-    {
-      case '\n':
-	file_recovery_new->extension=file_hint_ps.extension;
-	if(file_recovery_new->blocksize > 8)
-	  file_recovery_new->data_check=&data_check_ps;
-	return 1;
-      case 'E':
-	if(i+5 <= buffer_size && memcmp(&buffer[i],"EPSF-",5)==0)
-	{
-	  file_recovery_new->extension="eps";
-	  return 1;
-	}
-	break;
+  file_recovery_new->min_filesize = sizeof(ps_header);
+  file_recovery_new->file_check = &file_check_ps;
+
+  for (i = sizeof(ps_header); i < 20; i++) {
+    switch (buffer[i]) {
+    case '\n':
+      file_recovery_new->extension = file_hint_ps.extension;
+      if (file_recovery_new->blocksize > 8)
+        file_recovery_new->data_check = &data_check_ps;
+      return 1;
+    case 'E':
+      if ((unsigned int)i + 5 <= buffer_size && memcmp(&buffer[i], "EPSF-", 5) == 0) {
+        file_recovery_new->extension = "eps";
+        return 1;
+      }
+      break;
     }
   }
-  file_recovery_new->extension=file_hint_ps.extension;
-  if(file_recovery_new->blocksize > 8)
-    file_recovery_new->data_check=&data_check_ps;
+  file_recovery_new->extension = file_hint_ps.extension;
+  if (file_recovery_new->blocksize > 8)
+    file_recovery_new->data_check = &data_check_ps;
   return 1;
 }
 
-static void register_header_check_ps(file_stat_t *file_stat)
-{
-  register_header_check(0, ps_header,sizeof(ps_header), &header_check_ps, file_stat);
+static void register_header_check_ps(file_stat_t *file_stat) {
+  register_header_check(0, ps_header, sizeof(ps_header), &header_check_ps, file_stat);
 }
 #endif
